refactor(utils): Consolidate and rename export utility functions for timestamps and JSONL formatting
